refactor to extract functions for reuse
